WIP: Fixed most bugs but rainbow grades is not seeing submissions/scores correctly
