--- .setup/bin/make_customization_sample.py
@@ -1,212 +1,0 @@
-import pprint
-import yaml
-import os
-
-def load_data_yaml(file_path):
-    """
-    Loads yaml file from the .setup/data directory returning the parsed structure
-    :param file_path: name of file to load
-    :return: parsed YAML structure from loaded file
-    """
-    if not os.path.isfile(file_path):
-        raise IOError("Missing the yaml file {}".format(file_path))
-    with open(file_path) as open_file:
-        yaml_file = yaml.safe_load(open_file)
-    return yaml_file
-
-class Gradeable(object):
-    """
-    Attributes:
-        config_path
-        id
-        type
-    """
-    def __init__(self, gradeable):
-        self.id = ""
-        self.gradeable_config = None
-        self.config_path = None
-        self.sample_path = None
-        self.title = ""
-        self.instructions_url = ""
-        self.overall_ta_instructions = ""
-        self.team_assignment = False
-        self.grade_by_registration = True
-        self.is_repository = False
-        self.subdirectory = ""
-        self.use_ta_grading = True
-        self.late_days = 2
-        self.precision = 0.5
-        self.syllabus_bucket = "none (for practice only)"
-        self.min_grading_group = 3
-        self.grading_rotating = []
-        self.submissions = []
-        self.max_random_submissions = None
-
-        if 'gradeable_config' in gradeable:
-            self.gradeable_config = gradeable['gradeable_config']
-            self.type = 0
-
-            if 'g_id' in gradeable:
-                self.id = gradeable['g_id']
-            else:
-                self.id = gradeable['gradeable_config']
-
-            if 'eg_max_random_submissions' in gradeable:
-                self.max_random_submissions = int(gradeable['eg_max_random_submissions'])
-
-            if 'g_bucket' in gradeable:
-                self.syllabus_bucket = gradeable['g_bucket']
-
-            if 'config_path' in gradeable:
-                self.config_path = gradeable['config_path']
-            else:
-                examples_path = os.path.join(MORE_EXAMPLES_DIR, self.gradeable_config, "config")
-                tutorial_path = os.path.join(TUTORIAL_DIR, self.gradeable_config, "config")
-                if os.path.isdir(examples_path):
-                    self.config_path = examples_path
-                elif os.path.isdir(tutorial_path):
-                    self.config_path = tutorial_path
-                else:
-                    self.config_path = None
-
-            examples_path = os.path.join(MORE_EXAMPLES_DIR, self.gradeable_config, "submissions")
-            tutorial_path = os.path.join(TUTORIAL_DIR, self.gradeable_config, "submissions")
-            if 'sample_path' in gradeable:
-                self.sample_path = gradeable['sample_path']
-            else:
-                if os.path.isdir(examples_path):
-                    self.sample_path = examples_path
-                elif os.path.isdir(tutorial_path):
-                    self.sample_path = tutorial_path
-                else:
-                    self.sample_path = None
-        else:
-            self.id = gradeable['g_id']
-            self.type = int(gradeable['g_type'])
-            self.config_path = None
-            self.sample_path = None
-
-        assert 0 <= self.type <= 2
-
-        if 'g_title' in gradeable:
-            self.title = gradeable['g_title']
-        else:
-            self.title = self.id.replace("_", " ").title()
-
-        if 'g_grade_by_registration' in gradeable:
-            self.grade_by_registration = gradeable['g_grade_by_registration'] is True
-
-        if 'grading_rotating' in gradeable:
-            self.grading_rotating = gradeable['grading_rotating']
-
-        self.ta_view_date = parse_datetime(gradeable['g_ta_view_start_date'])
-        self.grade_start_date = parse_datetime(gradeable['g_grade_start_date'])
-        self.grade_released_date = parse_datetime(gradeable['g_grade_released_date'])
-        if self.type == 0:
-            self.submission_open_date = parse_datetime(gradeable['eg_submission_open_date'])
-            self.submission_due_date = parse_datetime(gradeable['eg_submission_due_date'])
-            if 'eg_is_repository' in gradeable:
-                self.is_repository = gradeable['eg_is_repository'] is True
-            if self.is_repository and 'eg_subdirectory' in gradeable:
-                self.subdirectory = gradeable['eg_subdirectory']
-            if 'eg_use_ta_grading' in gradeable:
-                self.use_ta_grading = gradeable['eg_use_ta_grading'] is True
-            if 'eg_late_days' in gradeable:
-                self.late_days = max(0, int(gradeable['eg_late_days']))
-            if 'eg_precision' in gradeable:
-                self.precision = float(gradeable['eg_precision'])
-            if self.config_path is None:
-                examples_path = os.path.join(MORE_EXAMPLES_DIR, self.id, "config")
-                tutorial_path = os.path.join(TUTORIAL_DIR, self.id, "config")
-                if os.path.isdir(examples_path):
-                    self.config_path = examples_path
-                elif os.path.isdir(tutorial_path):
-                    self.config_path = tutorial_path
-                else:
-                    self.config_path = None
-            assert self.ta_view_date < self.submission_open_date
-            assert self.submission_open_date < self.submission_due_date
-            assert self.submission_due_date < self.grade_start_date
-            if self.gradeable_config is not None:
-                if self.sample_path is not None:
-                    if os.path.isfile(os.path.join(self.sample_path, "submissions.yml")):
-                        self.submissions = load_data_yaml(os.path.join(self.sample_path, "submissions.yml"))
-                    else:
-                        self.submissions = os.listdir(self.sample_path)
-                        self.submissions = list(filter(lambda x: not x.startswith("."), self.submissions))
-                    if isinstance(self.submissions, list):
-                        for elem in self.submissions:
-                            if isinstance(elem, dict):
-                                raise TypeError("Cannot have dictionary inside of list for submissions "
-                                                "for {}".format(self.sample_path))
-        assert self.ta_view_date < self.grade_start_date
-        assert self.grade_start_date < self.grade_released_date
-
-        self.components = []
-        for i in range(len(gradeable['components'])):
-            component = gradeable['components'][i]
-            if self.type < 2:
-                component['gc_is_text'] = False
-            elif self.type > 0:
-                component['gc_ta_comment'] = ""
-                component['gc_student_comment'] = ""
-
-            if self.type == 1:
-                component['gc_max_value'] = 1
-            self.components.append(Component(component, i+1))
-
-class Component(object):
-    def __init__(self, component, order):
-        self.title = component['gc_title']
-        self.ta_comment = ""
-        self.student_comment = ""
-
-        self.is_text = False
-        self.is_extra_credit = False
-        self.order = order
-        if 'gc_ta_comment' in component:
-            self.ta_comment = component['gc_ta_comment']
-        if 'gc_student_comment' in component:
-            self.student_comment = component['gc_student_comment']
-        if 'gc_is_text' in component:
-            self.is_text = component['gc_is_text'] is True
-        if 'gc_is_extra_credit' in component:
-            self.is_extra_credit = component['gc_is_extra_credit'] is True
-
-        if self.is_text:
-            self.max_value = 0
-        else:
-            self.max_value = float(component['gc_max_value'])
-
-        self.key = None
-
-    def create(self, g_id, conn, table):
-        ins = table.insert().values(g_id=g_id, gc_title=self.title, gc_ta_comment=self.ta_comment,
-                                    gc_student_comment=self.student_comment,
-                                    gc_max_value=self.max_value, gc_is_text=self.is_text,
-                                    gc_is_extra_credit=self.is_extra_credit, gc_order=self.order)
-        res = conn.execute(ins)
-        self.key = res.inserted_primary_key[0]
-
-def main():
-    course_file = os.path.join("..","data","courses","sample.yml")
-    course_json = load_data_yaml(course_file)
-
-    course_id = course_json['code']
-    print("Making customization.json for course: {}".format(course_id))
-
-    gradeables = {}
-
-    for g in course_json['gradeables']:
-        gradeable = Gradeable(g)
-        if gradeable.syllabus_bucket not in gradeables:
-            gradeables[gradeable.syllabus_bucket] = []
-        gradeables[gradeable.syllabus_bucket].append(gradeable)
-
-    for bucket,g_list in gradeables.items():
-        print("There are {} gradeables in category {}".format(len(g_list),bucket))
-
-    # pprint.pprint(course_json)
-
-if __name__ == '__main__':
-    main()
