+++ tests/rainbowGrades/customization_sample.json
@@ -1,76 +1,53 @@
 {
-    "display": [
-        "instructor_notes",
-        "grade_summary",
-        "grade_details"
-        //"exam_seating",
-        //"moss_details",
-        //"final_grade",
-        //"iclicker"
-    ],
-    "display_benchmark": [
-        "average",
-        "stddev",
-        "perfect",
-        "lowest_a-",
-        "lowest_b-",
-        "lowest_c-",
-        "lowest_d"
-    ],
-    // These are the default curves applied to each benchmark
-    // (can be overridden per item below)
-    "benchmark_percent": {
-	"lowest_a-": 0.9,
-	"lowest_b-": 0.8,
-	"lowest_c-": 0.7,
-	"lowest_d": 0.6
-    },
-    "gradeables": [
-	{ "type" : "hw",
-	    "count": 4,
-	    "percent": 1.00,
-	    "ids": [
-		{ "id":"python_static_analysis", "max": 14, "curve": [4.50, 4.00, 3.50, 3.00]},
-		{ "id":"03_multipart", "max": 15, "curve": [14.50, 14.00, 13.50, 13.00] },
-		{ "id":"python_linehighlight", "max": 95, "curve": [94.50, 64.00, 23.50, 13.00] },
-		{ "id":"python_simple_homework", "max": 5, "curve": [4.0, 4.0, 3.0, 3.0] }
-	    ]
-	}
-    ],
-
-/*,
-   //"final_cutoff": {
-    //	"A": 93.0,
-    //	"A-": 90.0,
-    //	"B+": 87.0,
-    //	"B": 83.0,
-    //	"B-": 80.0,
-    //	"C+": 77.0,
-    //	"C": 73.0,
-    //	"C-": 70.0,
-    //	"D+": 65.0,
-    //	"D": 60.0
-    //},
-*/
-
-    "section": {
-	"1": "TA_name_1",
-	"2": "TA_name_2",
-	"3": "TA_name_3",
-	"4": "TA_name_4",
-	"5": "TA_name_1",
-	"6": "TA_name_2",
-	"7": "TA_name_3",
-	"8": "TA_name_4",
-  "9": "TA_name_5",
-	"10": "TA_name_5"
-    },
-    "messages": [
-	"<b>My Favorite CS Class</b>",
-	"Note: Please be patient with data entry/grade corrections for the most recent lab, homework, and test.",
-	"Please contact your graduate lab TA if a grade remains missing or incorrect for more than a week."
-	//"<em>Note: All grades for lab & hw are uploaded.</em>",
-	//"<em>Note: All regrade requests have been processed.</em>"
-    ]
+  "display": [
+      "instructor_notes",
+      "grade_summary",
+      "grade_details"
+  ],
+  "display_benchmark": [
+      "average",
+      "stddev",
+      "perfect",
+      "lowest_a-",
+      "lowest_b-",
+      "lowest_c-",
+      "lowest_d"
+  ],
+  "benchmark_percent": {
+    "lowest_a-": 0.9,
+    "lowest_b-": 0.8,
+    "lowest_c-": 0.7,
+    "lowest_d": 0.6
+  },
+  "gradeables": [
+    {
+      "type" : "hw",
+      "count": 4,
+      "percent": 1.00,
+      "ids": [
+        { "id":"python_static_analysis", "max": 14, "curve": [4.50, 4.00, 3.50, 3.00]},
+        { "id":"03_multipart", "max": 15, "curve": [14.50, 14.00, 13.50, 13.00] },
+        { "id":"python_linehighlight", "max": 95, "curve": [94.50, 64.00, 23.50, 13.00] },
+        { "id":"python_simple_homework", "max": 5, "curve": [4.0, 4.0, 3.0, 3.0] }
+      ]
+    }
+  ],
+  "section": {
+    "1": "TA_name_1",
+    "2": "TA_name_2",
+    "3": "TA_name_3",
+    "4": "TA_name_4",
+    "5": "TA_name_1",
+    "6": "TA_name_2",
+    "7": "TA_name_3",
+    "8": "TA_name_4",
+    "9": "TA_name_5",
+    "10": "TA_name_5"
+  },
+  "messages": [
+    "<b>My Favorite CS Class</b>",
+    "Note: Please be patient with data entry/grade corrections for the most recent lab, homework, and test.",
+    "Please contact your graduate lab TA if a grade remains missing or incorrect for more than a week."
+  ]
 }
